Use dedicated DBs for e2e tests
