+++ test/jest-e2e-tests/accessGroups.e2e-spec.ts
@@ -1,9 +1,8 @@
 import request from "supertest";
-import { Test } from "@nestjs/testing";
 import { INestApplication } from "@nestjs/common";
-import { AppModule } from "src/app.module";
 import { UsersService } from "src/users/users.service";
 import { User } from "src/users/schemas/user.schema";
+import { createTestingModuleFactory } from "./utlis";
 
 describe("Access groups test", () => {
   let app: INestApplication;
@@ -11,9 +10,7 @@ describe("Access groups test", () => {
   let u: User;
 
   beforeAll(async () => {
-    const moduleFixture = await Test.createTestingModule({
-      imports: [AppModule],
-    }).compile();
+    const moduleFixture = await createTestingModuleFactory().compile();
 
     app = moduleFixture.createNestApplication();
     app.setGlobalPrefix("api/v3");
